requirements.txt
Flask
openai
requests
python-dotenv
azure-cognitiveservices-speech
